requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
boto3==1.35.69
botocore==1.35.69
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
dj-database-url==2.3.0
Django==4.2.16
django-extensions==3.2.3
django-rest-passwordreset==1.5.0
django-storages==1.14.4
djangorestframework==3.15.2
drf-extensions==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
idna==3.10
iniconfig==2.0.0
ipython==8.29.0
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mock==5.1.0
netifaces==0.11.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
polib==1.2.0
prettytable==3.12.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pytest==8.3.3
pytest-django==4.9.0
pytest-env==1.1.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
ruff==0.8.0
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
tomli==2.1.0
traitlets==5.14.3
typing_extensions==4.12.2
ua-parser==0.18.0
Unidecode==1.3.8
urllib3==2.2.3
user-agents==2.2.0
wcwidth==0.2.13
Werkzeug==3.1.3
XlsxWriter==3.2.0
uwsgi==2.0.24
psycopg==3.1.18
